Monte Hall function that requires user inputs (R)
monte_hall is a funtion where the user passed a set number of doors.
The function then prompts the user to pick a door form the set.  An evil
genious removes doors 1 by 1 until there are 2 left.  The door the
player chose and one other.  The player then inputs which door they
would like to reveal and the function returns either a Win or Loss.